# docs/apifox-full.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuraStack 接口文档（自动生成）",
    "version": "1.0.0",
    "description": "AuraStack 接口文档（由脚本自动生成）。"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Development"
    },
    {
      "url": "http://localhost:5000",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Flask Session Cookie 登录态"
      }
    }
  },
  "paths": {
...
    "/api/admin/roles": {
      "get": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "查询roles",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "创建roles",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          },
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/roles/{role_id}": {
      "delete": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "删除roles",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：role_id"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "更新roles",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：role_id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/admin/roles/export": {
      "post": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "导出roles",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/admin/roles/template": {
      "get": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "下载roles导入模板",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/roles/import": {
      "post": {
        "tags": [
          "后台-角色管理"
        ],
        "summary": "导入roles",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/角色管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/admin/scheduled-tasks": {
      "get": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "查询scheduled-tasks",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "创建scheduled-task",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          },
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/scheduled-tasks/{task_id}": {
      "get": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "查询scheduled-task",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：task_id"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "更新scheduled-task",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：task_id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "删除scheduled-task",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：task_id"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/scheduled-tasks/{task_id}/run": {
      "post": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "执行scheduled-task",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：task_id"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/scheduled-tasks/runs": {
      "get": {
        "tags": [
          "后台-定时任务"
        ],
        "summary": "查询scheduled-task-runs",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/定时任务",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    }
  }
}